fb-common/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "fb-common"
version = "0.1.0"
description = "Shared env / MySQL SQLAlchemy helpers for fb services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pymysql>=1.1.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
